+++ src/main/java/net/k40s/FileUtils.java
@@ -2,7 +2,6 @@ package net.k40s;
 
 import java.io.File;
 import java.io.FileInputStream;
-import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipOutputStream;
