+++ src/main/java/net/k40s/HomePage.java
@@ -1,15 +1,13 @@
 package net.k40s;
 
 import net.k40s.album.AlbumsPage;
-import net.k40s.debug.DBTestSite;
 import net.k40s.single.SinglesPage;
 
+import org.apache.wicket.markup.html.WebPage;
 import org.apache.wicket.markup.html.basic.Label;
 import org.apache.wicket.markup.html.link.ExternalLink;
 import org.apache.wicket.markup.html.link.Link;
 import org.apache.wicket.request.mapper.parameter.PageParameters;
-import org.apache.wicket.markup.html.WebMarkupContainer;
-import org.apache.wicket.markup.html.WebPage;
 
 public class HomePage extends WebPage {
 
@@ -39,19 +37,7 @@ public class HomePage extends WebPage {
 				setResponsePage(AlbumsPage.class);
 			}
 		});
-		WebMarkupContainer debugListItem = new WebMarkupContainer("debugLi");
-		debugListItem.add(new Link("debugLink"){
-			@Override
-			public void onClick(){
-				setResponsePage(DBTestSite.class);
-			}
-		});
-		if(MusicSite.isDebug()){
-			add(debugListItem);
-		} else {
-			debugListItem.setVisible(false);
-			add(debugListItem);
-		}
+		
 		add(new ExternalLink("contactLink", "http://k40s.net"));
 		add(new Label("version", Storage.getVersion()));
 	}
