+++ src/main/java/net/k40s/backend/SignInPage.java
@@ -3,6 +3,7 @@ package net.k40s.backend;
 import net.k40s.HomePage;
 import net.k40s.album.AlbumsPage;
 import net.k40s.single.SinglesPage;
+
 import org.apache.wicket.authroles.authentication.AuthenticatedWebSession;
 import org.apache.wicket.markup.html.WebPage;
 import org.apache.wicket.markup.html.form.PasswordTextField;
