--- src/main/java/net/k40s/debug/DBTestSite.html
@@ -1,80 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:wicket="http://wicket.apache.org">
-<head>
-	<meta charset="utf-8">
-	<meta http-equiv="X-UA-Compatible" content="IE=edge">
-	<meta name="viewport" content="width=device-width, initial-scale=1">
-	<title>K40S Music</title>
-
-	<!-- Bootstrap -->
-	<link href="css/bootstrap.min.css" rel="stylesheet">
-	<!--link href="css/bootstrap-theme.min.css" rel="stylesheet"-->
-	<link href="css/material-wfont.min.css" rel="stylesheet">
-	<link href="css/ripples.min.css" rel="stylesheet">
-	<link href="css/style.css" rel="stylesheet">
-	<link rel="shortcut icon" href="img/favicon.ico" type="image/x-icon">
-</head>
-<body>
-<script type="text/javascript">
-	(function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
-		(i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
-			m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
-	})(window,document,'script','//www.google-analytics.com/analytics.js','ga');
-
-	ga('create', 'UA-44457715-5', 'auto');
-	ga('send', 'pageview');
-
-</script>
-<nav class="navbar navbar-inverse navbar-fixed-top">
-	<div class="container">
-		<div class="navbar-header">
-			<button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#navbar" aria-expanded="false" aria-controls="navbar">
-				<span class="sr-only">Toggle navigation</span>
-				<span class="icon-bar"></span>
-				<span class="icon-bar"></span>
-				<span class="icon-bar"></span>
-			</button>
-			<a class="navbar-brand" href="#" wicket:id="homeLink">K40S</a>
-		</div>
-		<div id="navbar" class="collapse navbar-collapse">
-			<ul class="nav navbar-nav">
-				<li ><a href="#" wicket:id="singlesLink">Singles</a></li>
-				<li ><a href="#" wicket:id="albumsLink">Albums</a></li>
-				<li><a href="#" wicket:id="contactLink">Contact</a></li>
-				<li class="active" wicket:id="debugLi"><a href="#" wicket:id="debugLink">Debug</a></li>
-			</ul>
-			<p class="navbar-text navbar-right" id="version" style="display:none">Beta <span wicket:id="version"></span></p>
-		</div><!--/.nav-collapse -->
-	</div>
-</nav>
-
-
-<div class="container">
-<div class="panel panel-default">
-<div class="panel-heading">
-<h1 class="panel-title">Welcome</h1>
-</div>
-<div class="panel-body">
-<p><b>This</b> is the debug site. You should only see this, when the application runs in development mode.</p>
-<hr>
-<p><b>The database</b> debug page is located <a href="#" wicket:id="dbDebugLink">here</a>.</p>
-</div>
-</div>
-</div>
-
-<!-- jQuery (necessary for Bootstrap's JavaScript plugins) -->
-<script src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.1/jquery.min.js"></script>
-<!-- Include all compiled plugins (below), or include individual files as needed -->
-<script src="js/bootstrap.min.js"></script>
-<script src="js/jquery-ui.min.js"></script>
-<script src="js/material.min.js"></script>
-<script src="js/ripples.min.js"></script>
-<script src="js/version.js"></script>
-<script>
-    $(document).ready(function() {
-        $.material.init();
-    });
-</script>
-
-</body>
-</html>
