+++ src/main/java/net/k40s/single/SingleProductPage.java
@@ -1,12 +1,14 @@
 package net.k40s.single;
 
+import java.io.File;
+import java.io.Serializable;
+import java.util.ArrayList;
+import java.util.List;
+
 import net.k40s.HomePage;
-import net.k40s.MusicSite;
 import net.k40s.Storage;
 import net.k40s.album.AlbumsPage;
-import net.k40s.debug.DBTestSite;
 
-import org.apache.wicket.markup.html.WebMarkupContainer;
 import org.apache.wicket.markup.html.WebPage;
 import org.apache.wicket.markup.html.basic.Label;
 import org.apache.wicket.markup.html.image.Image;
@@ -19,11 +21,6 @@ import org.apache.wicket.request.mapper.parameter.PageParameters;
 import org.apache.wicket.request.resource.ContextRelativeResource;
 import org.wicketstuff.html5.media.MediaSource;
 import org.wicketstuff.html5.media.audio.Html5Audio;
-
-import java.io.File;
-import java.io.Serializable;
-import java.util.ArrayList;
-import java.util.List;
 
 /**
  * @author Lukas F&uuml;lling (l.fuelling@micromata.de)
@@ -99,18 +96,6 @@ public class SingleProductPage extends WebPage implements Serializable {
     });
     add(new Label("version", Storage.getVersion()));
     add(new ExternalLink("contactLink", "http://k40s.net"));
-    WebMarkupContainer debugListItem = new WebMarkupContainer("debugLi");
-	debugListItem.add(new Link("debugLink"){
-		@Override
-		public void onClick(){
-			setResponsePage(DBTestSite.class);
-		}
-	});
-	if(MusicSite.isDebug()){
-		add(debugListItem);
-	} else {
-		debugListItem.setVisible(false);
-		add(debugListItem);
-	}
+    
   }
 }
